split pro routes
