+++ lib/karafka/web/pro/ui/routes/dashboard.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+# This Karafka component is a Pro component under a commercial license.
+# This Karafka component is NOT licensed under LGPL.
+#
+# All of the commercial components are present in the lib/karafka/pro directory of this
+# repository and their usage requires commercial license agreement.
+#
+# Karafka has also commercial-friendly license, commercial support and commercial components.
+#
+# By sending a pull request to the pro components, you are agreeing to transfer the copyright of
+# your code to Maciej Mensfeld.
+
+module Karafka
+  module Web
+    module Pro
+      module Ui
+        module Routes
+          # Manages the dashboard related routes
+          class Dashboard < Base
+            route do |r|
+              r.get 'dashboard' do
+                @breadcrumbs = false
+                controller = Controllers::DashboardController.new(params)
+                controller.index
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end
