+++ lib/karafka/web/pro/ui/routes/messages.rb
@@ -1,0 +1,43 @@
+# frozen_string_literal: true
+
+# This Karafka component is a Pro component under a commercial license.
+# This Karafka component is NOT licensed under LGPL.
+#
+# All of the commercial components are present in the lib/karafka/pro directory of this
+# repository and their usage requires commercial license agreement.
+#
+# Karafka has also commercial-friendly license, commercial support and commercial components.
+#
+# By sending a pull request to the pro components, you are agreeing to transfer the copyright of
+# your code to Maciej Mensfeld.
+
+module Karafka
+  module Web
+    module Pro
+      module Ui
+        module Routes
+          # Manages the Kafka messages related routes
+          class Messages < Base
+            route do |r|
+              r.on 'messages' do
+                controller = Controllers::MessagesController.new(params)
+
+                r.post String, Integer, Integer, 'republish' do |topic_id, partition_id, offset|
+                  controller.republish(topic_id, partition_id, offset)
+                end
+
+                r.get String, Integer, Integer, 'download' do |topic_id, partition_id, offset|
+                  controller.download(topic_id, partition_id, offset)
+                end
+
+                r.get String, Integer, Integer, 'export' do |topic_id, partition_id, offset|
+                  controller.export(topic_id, partition_id, offset)
+                end
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end
