+++ lib/karafka/web/pro/ui/routes/recurring_tasks.rb
@@ -1,0 +1,67 @@
+# frozen_string_literal: true
+
+# This Karafka component is a Pro component under a commercial license.
+# This Karafka component is NOT licensed under LGPL.
+#
+# All of the commercial components are present in the lib/karafka/pro directory of this
+# repository and their usage requires commercial license agreement.
+#
+# Karafka has also commercial-friendly license, commercial support and commercial components.
+#
+# By sending a pull request to the pro components, you are agreeing to transfer the copyright of
+# your code to Maciej Mensfeld.
+
+module Karafka
+  module Web
+    module Pro
+      module Ui
+        module Routes
+          # Manages the recurring tasks feature related routes
+          class RecurringTasks < Base
+            route do |r|
+              r.on 'recurring_tasks' do
+                controller = Controllers::RecurringTasksController.new(params)
+
+                r.get 'schedule' do
+                  controller.schedule
+                end
+
+                r.get 'logs' do
+                  controller.logs
+                end
+
+                r.post 'trigger_all' do
+                  controller.trigger_all
+                end
+
+                r.post 'enable_all' do
+                  controller.enable_all
+                end
+
+                r.post 'disable_all' do
+                  controller.disable_all
+                end
+
+                r.post String, 'trigger' do |task_id|
+                  controller.trigger(task_id)
+                end
+
+                r.post String, 'enable' do |task_id|
+                  controller.enable(task_id)
+                end
+
+                r.post String, 'disable' do |task_id|
+                  controller.disable(task_id)
+                end
+
+                r.get do
+                  r.redirect root_path('recurring_tasks/schedule')
+                end
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end
